E_NURSE/backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.0.1
python-dateutil==2.8.2
flask-restx==1.1.0
gunicorn==21.2.0
pytest==7.4.2
flask-jwt-extended==4.5.2 